requirements.txt
Django==3.1.4
djangorestframework==3.12.2
jwt==1.1.0
Markdown==3.3.3
